Added auth client and fixed docs
