+++ lib/src/core/index.ts
@@ -15,5 +15,4 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-export * from "./client";
+export * from "./authentication";
